refactor(): Players became an use case because there is an action in Players class
